Add option to specify date/time labels for N days into future
By default, the script will now create labels at 7am and 9pm for each of the next eight days, and then preserve those labels even if they are empty, cleaning up after itself as time passes. This is intended to solve the problem of a) having to create new labels all the time, and b) not being able to use relative labels of 'tonight' since often I want to wait days before dealing with an email.

And a tiny bit of whitespace cleanup.